requirements.txt
streamlit
pandas
numpy
plotly
mysql-connector-python
folium
streamlit-folium
branca
python-dotenv